chore: 杂项修改。 fix: 修复已知bug，例如登记页面显示错误bug。

+++ backend/src/main/java/com/caixy/adminSystem/model/dto/StudentGrades/StudentGradesAddRequest.java
@@ -57,7 +57,7 @@ public class StudentGradesAddRequest implements Serializable
          * 成绩
          */
         @NotNull(message = "成绩不能为空")
-        private Long totalGrade;
+        private Integer totalGrade;
 
         /**
          * 平时成绩
